Move config => api/config
