+++ cli/server/proxy/proxy.go
@@ -21,13 +21,13 @@ import (
 	"sync"
 
 	"github.com/docker/compose-cli/api/client"
+	"github.com/docker/compose-cli/api/config"
 	composev1 "github.com/docker/compose-cli/cli/server/protos/compose/v1"
 	containersv1 "github.com/docker/compose-cli/cli/server/protos/containers/v1"
 	contextsv1 "github.com/docker/compose-cli/cli/server/protos/contexts/v1"
 	streamsv1 "github.com/docker/compose-cli/cli/server/protos/streams/v1"
 	volumesv1 "github.com/docker/compose-cli/cli/server/protos/volumes/v1"
 	"github.com/docker/compose-cli/cli/server/proxy/streams"
-	"github.com/docker/compose-cli/config"
 )
 
 type clientKey struct{}
