+++ local/compose/push.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/docker/buildx/driver"
 
+	"github.com/docker/compose-cli/api/config"
 	"github.com/docker/compose-cli/api/progress"
-	"github.com/docker/compose-cli/config"
 
 	"github.com/compose-spec/compose-go/types"
 	cliconfig "github.com/docker/cli/cli/config"
